requirements.txt
peewee~=3.16.0
setuptools~=63.2.0
certifi~=2022.9.24
requests~=2.28.1
idna~=3.4
urllib3~=1.26.13
python-dotenv~=0.21.0
pyTelegramBotAPI~=4.9.0
python-telegram-bot-calendar~=1.0.5
loguru~=0.7.0

six~=1.16.0